requirements.txt
openai>=1.30.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
chromadb>=0.5.0
redis>=5.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0
